requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google Gemini AI
google-generativeai==0.3.2

# Vector Database (Pinecone)
pinecone-client==2.2.4

# Embedding Models
sentence-transformers==2.2.2

# LangChain Framework
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.6
langchain-pinecone==0.0.3
langchain-embeddings-huggingface==0.0.1

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Data Validation
pydantic==2.5.0

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Text Processing
nltk==3.8.1
regex==2023.10.3

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# CORS Support
fastapi-cors

# Logging
loguru==0.7.2

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4

# Optional: For better JSON handling
orjson==3.9.10

# Optional: For monitoring and metrics
prometheus-client==0.19.0

# Optional: For caching
redis==5.0.1

# Development and Testing (optional - comment out for production)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2